pyproject.toml
[project]
name = "dreamstore"
version = "0.1.0"
description = "Project Academic about DialogFlow"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.2",
    "python-decouple>=3.8",
    "django-extensions>=3.2.3",
    "ipython>=8.29.0",
    "dj-database-url>=2.3.0",
    "django-cors-headers>=4.6.0",
    "drf-yasg>=1.21.8",
    "django-filter>=24.3",
    "requests>=2.32.3",
    "djangorestframework>=3.15.2",
    "psycopg2-binary>=2.9.10",
    "whitenoise>=6.8.2",
    "ruff>=0.7.2",
]
[tool.uv]
cache-dir = "./.uv_cache"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".pyproject.toml",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vercel",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "store/migrations",
    "staticfiles",
]

# Same as Black.
line-length = 120
indent-width = 4

# Assume Python 3.12
target-version = "py312"

lint.select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "UP",  # pyupgrade,
  "I",   # isort
]
ignore = [
    "E501",
    "E741"
]